% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diffusion.R
\name{Diffusion_net_analysis}
\alias{Diffusion_net_analysis}
\title{Title}
\usage{
Diffusion_net_analysis(
  SC_obj,
  Spot_manifest,
  savePath,
  centre = 1,
  broad = 1,
  spaceFile,
  Maskfile,
  imagefile,
  spot_r_min = 12,
  img_import = 1,
  spot_r_max = 20,
  Step = 10,
  diffusion_rate = 0.1,
  netfile,
  cluster_num = 50,
  geneenrichment = F,
  genelist = NULL
)
}
\arguments{
\item{SC_obj}{You can also load RNA-seq use an existed Seruat object.}

\item{Spot_manifest}{The TIST results returned by function "Meta_St_img_unsupervised"}

\item{savePath}{The address where the results stored.}

\item{centre}{The covariance matrix of distance Gaussian kernel, we default element in covariance matrix is identical.}

\item{broad}{The mean of distance Gaussian kernel.}

\item{spaceFile}{The address where the local file of all spots is stored. Such as "spatial/tissue_positions_list.csv".}

\item{Maskfile}{The address where the Mask file is stored. Such as "/home/data/Imginit/mask.txt". You can use no mask to generate an all 1 matrix. The more recommended approach is generate this mask file use our python code in this package.}

\item{imagefile}{The address where the Image file is stored. Such as "/home/data/spatial/tissue_hires_image.png".}

\item{spot_r_min}{A min window to select Image features. Default is 12.}

\item{img_import}{The image feature weight for feature integration.}

\item{spot_r_max}{A max window to select Image features. Default is 20.}

\item{Step}{Control the step length of random walk for walkstrap method.}

\item{diffusion_rate}{The diffusion rate, is related to temperature, liquid viscosity, and molecular size.}

\item{netfile}{The TIST-net build by function "Meta_St_img_unsupervised"}

\item{cluster_num}{The number of init cluster for image. Default is 50.}

\item{geneenrichment}{Whether to perform gene enhancement analysis. Default True.}

\item{genelist}{if geneenrichment is True, the analysis gene list.}
}
\description{
Title
}
\examples{
\donttest{
    Diffusion_net_analysis(SC_obj=SC_obj,
    Spot_manifest = Spot_manifest_imgunsup,
    savePath= savePath,
    centre = 1,
    broad = 2,
    spaceFile=spaceFile,
    Maskfile=Maskfile,
    imagefile=imagefile,
    spot_r_min = 12,
    img_import = 1,
    spot_r_max = 20,
    Step = NULL,
    diffusion_rate = dd,
    netfile = netfile,
    geneenrichment = T,
    genelist = NULL)
}

}
