% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Res_net.R
\name{Dropout_net_analysis}
\alias{Dropout_net_analysis}
\title{Title}
\usage{
Dropout_net_analysis(
  SC_obj,
  Spot_manifest,
  savePath,
  spaceFile,
  Maskfile,
  imagefile,
  spot_r_min = 12,
  img_import = 1,
  spot_r_max = 20,
  Step = 10,
  dropout_rate = 0.1,
  netfile,
  cluster_num = 50,
  geneenrichment = F,
  genelist = NULL
)
}
\arguments{
\item{SC_obj}{You can also load RNA-seq use an existed Seruat object.}

\item{Spot_manifest}{The TIST results returned by function "Meta_St_img_unsupervised"}

\item{savePath}{The address where the results stored.}

\item{spaceFile}{The address where the local file of all spots is stored. Such as "spatial/tissue_positions_list.csv".}

\item{Maskfile}{The address where the Mask file is stored. Such as "/home/data/Imginit/mask.txt". You can use no mask to generate an all 1 matrix. The more recommended approach is generate this mask file use our python code in this package.}

\item{imagefile}{The address where the Image file is stored. Such as "/home/data/spatial/tissue_hires_image.png".}

\item{spot_r_min}{A min window to select Image features. Default is 12.}

\item{img_import}{The image feature weight for feature integration.}

\item{spot_r_max}{A max window to select Image features. Default is 20.}

\item{Step}{Control the step length of random walk for walkstrap method.}

\item{dropout_rate}{The dropout rate.}

\item{netfile}{The TIST-net build by function "Meta_St_img_unsupervised"}

\item{cluster_num}{The number of init cluster for image. Default is 50.}

\item{geneenrichment}{Whether to perform gene enhancement analysis. Default True.}

\item{genelist}{if geneenrichment is True, the analysis gene list.}
}
\description{
Title
}
\examples{
\donttest{
    Dropout_net_analysis(SC_obj = SC_obj,
    Spot_manifest = Spot_manifest_imgunsup,
    savePath = savePath,
    spaceFile =spaceFile,
    Maskfile = Maskfile,
    imagefile = imagefile,
    spot_r_min = 12,
    img_import = 1,
    spot_r_max = 20,
    dropout_rate = dd,
    netfile = netfile,
    geneenrichment = T,
    genelist = NULL)


}
}
