% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_plot.R
\name{Meta_expr_matrix}
\alias{Meta_expr_matrix}
\title{Title}
\usage{
Meta_expr_matrix(
  exprPath,
  Spot_manifest,
  savePath,
  imagefile,
  merge_method = "max",
  method = "Walktrap_id"
)
}
\arguments{
\item{exprPath}{The address where the RNA-seq file is stored. Such as "/home/data/filtered_feature_bc_matrix/".}

\item{Spot_manifest}{The TIST results returned by function "Meta_St_img_unsupervised"}

\item{savePath}{The address where the results stored.}

\item{imagefile}{The address where the Image file is stored. Such as "/home/data/spatial/tissue_hires_image.png".}

\item{merge_method}{mean or max for neighbor expression}

\item{method}{TIST network method is "Walktrap_id"}
}
\value{
Mc_manifest
}
\description{
Title
}
