% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meta_ST_img.R
\name{Meta_St_img_unsupervised}
\alias{Meta_St_img_unsupervised}
\title{Title}
\usage{
Meta_St_img_unsupervised(
  Maskfile = NULL,
  imagefile,
  exprPath,
  SC_obj = NULL,
  spaceFile,
  colors = NULL,
  savePath,
  Method = "walktrap",
  imgMethod = "Marcov",
  imginitMethod = "Kmeans",
  Step = NULL,
  spot_r_min = 12,
  spot_r_max = 20,
  cluster_num = 50,
  maxiter = 10,
  sacle_score = 0.09920635,
  img_import = 1
)
}
\arguments{
\item{Maskfile}{The address where the Mask file is stored. Such as "/home/data/Imginit/mask.txt". You can use no mask to generate an all 1 matrix. The more recommended approach is generate this mask file use our python code in this package.}

\item{imagefile}{The address where the Image file is stored. Such as "/home/data/spatial/tissue_hires_image.png".}

\item{exprPath}{The address where the RNA-seq file is stored. Such as "/home/data/filtered_feature_bc_matrix/".}

\item{SC_obj}{If exprPath is NULL, you can also load RNA-seq use an existed Seruat object. Default is NULL.}

\item{spaceFile}{The address where the local file of all spots is stored. Such as "spatial/tissue_positions_list.csv".}

\item{colors}{Color bar for visualization. We will give a default color scheme if NULL.}

\item{savePath}{The address where the results stored.}

\item{Method}{The Network-based methods to find clusters. Default param is walktrap,}

\item{imgMethod}{The image feather select method. Default is Marcov, Other optional: Kmeans. We will use original image feature if NULL.}

\item{imginitMethod}{If imgMethod is Marcov, this param is to select label init method, such as "Kmeans" default or "Random".}

\item{Step}{Control the step length of random walk for walkstrap method.}

\item{spot_r_min}{A min window to select Image features. Default is 12.}

\item{spot_r_max}{A max window to select Image features. Default is 20.}

\item{cluster_num}{The number of init cluster for image. Default is 50.}

\item{maxiter}{The number of iterations for marcov. Default is 10.}

\item{sacle_score}{The scale to map the spots coordinate in spaceFile with image.}

\item{img_import}{The image feature weight for feature integration.}
}
\value{
TIST results.
}
\description{
Title
}
\examples{
\donttest{
  Spot_manifest_imgunsup <- Meta_St_img_unsupervised(Maskfile = Maskfile,
  imagefile = imagefile,
  spaceFile = spaceFile,
  exprPath = exprPath,
  colors = NULL,
  savePath = savePath,
  Method = "walktrap",
  sacle_score = sacle_score)
}


}
