% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaceDiffGene.R
\name{SpaceDiffGene}
\alias{SpaceDiffGene}
\title{Title}
\usage{
SpaceDiffGene(
  SC_obj,
  Spot_manifest,
  savePath,
  MC_obj_file,
  Mc_manifest,
  Mc_pair_SC,
  methods = "walktrap",
  Mc_name,
  SPARK_file,
  MC_ident_file = NULL,
  DEgeneplot = T,
  Score_show = T,
  spark_methods = "SPARKX",
  netfile = NULL
)
}
\arguments{
\item{SC_obj}{load RNA-seq use an existed Seruat object}

\item{Spot_manifest}{The TIST results returned by function "Meta_St_img_unsupervised"}

\item{savePath}{The address where the results stored.}

\item{MC_obj_file}{MC obj}

\item{Mc_manifest}{Mc_manifest}

\item{Mc_pair_SC}{Mc_pair_SC}

\item{methods}{TIST network method is "Walktrap_id"}

\item{Mc_name}{Mc_name}

\item{SPARK_file}{SPARK_file generate by SPARK}

\item{MC_ident_file}{MC_ident_file}

\item{DEgeneplot}{DEgeneplot}

\item{Score_show}{T/F}

\item{spark_methods}{SPARK/SPARK-x}

\item{netfile}{The TIST-net build by function "Meta_St_img_unsupervised"}
}
\value{
MD_markers
}
\description{
Title
}
