% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARK.R
\name{Spark_methods}
\alias{Spark_methods}
\title{Title}
\usage{
Spark_methods(exprPath, spaceFile, savePath)
}
\arguments{
\item{exprPath}{The address where the RNA-seq file is stored. Such as "/home/data/filtered_feature_bc_matrix/".}

\item{spaceFile}{The address where the local file of all spots is stored. Such as "spatial/tissue_positions_list.csv".}

\item{savePath}{The address where the results stored.}
}
\description{
Title
}
\examples{
\donttest{
   Spark_methods(exprPath,
   spaceFile,
   savePath)
}


}
